task 2
Class Department extends the properties of Employee. Employee has method for setting
the values for employee name, number and a method for displaying the same. Similarly
department has method for setting the values for department name, number and a method
for displaying the same. Write a java program for above problem statement using single
inheritance.